moved colormap picker to VectorLegendAdvanced
